requirements.txt
Django==2.1.7
django-cors-headers==2.4.0
django-filter==2.1.0
djangorestframework==3.9.1
Markdown==3.0.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
python-decouple==3.1
pytz==2018.9
